chore: separate trait from field implementation
