+++ acvm-repo/acir_field/src/lib.rs
@@ -5,19 +5,21 @@
 
 use num_bigint::BigUint;
 use num_traits::Num;
+
+mod field_element;
 mod generic_ark;
 
 pub use generic_ark::AcirField;
 
 /// Temporarily exported generic field to aid migration to `AcirField`
-pub use generic_ark::FieldElement as GenericFieldElement;
+pub use field_element::FieldElement as GenericFieldElement;
 
 cfg_if::cfg_if! {
     if #[cfg(feature = "bls12_381")] {
-        pub type FieldElement = generic_ark::FieldElement<ark_bls12_381::Fr>;
+        pub type FieldElement = field_element::FieldElement<ark_bls12_381::Fr>;
         pub const CHOSEN_FIELD : FieldOptions = FieldOptions::BLS12_381;
     } else {
-        pub type FieldElement = generic_ark::FieldElement<ark_bn254::Fr>;
+        pub type FieldElement = field_element::FieldElement<ark_bn254::Fr>;
         pub const CHOSEN_FIELD : FieldOptions = FieldOptions::BN254;
     }
 }
